+++ Sources/Models/InstagramClient.swift
@@ -12,6 +12,10 @@ struct InstagramClient {
     let redirectURI: String?
     var scopes: [InstagramScope] = [.basic]
 
+    var stringScopes: String {
+        return scopes.map({ "\($0.rawValue)" }).joined(separator: "+")
+    }
+
     init(clientId: String?, redirectURI: String?) {
         self.clientId = clientId
         self.redirectURI = redirectURI